+++ common/i18n/src/test/java/is/codion/common/i18n/MessagesTest.java
@@ -49,5 +49,6 @@ public class MessagesTest {
 		Messages.saveMnemonic();
 		Messages.open();
 		Messages.openMnemonic();
+		Messages.filterVerb();
 	}
 }